This patch adds a safety automation to re-center the steering wheel when the car is parked and powered off
This is an enhanced version of controlsd.py from OpenPilot, which governs core control logic of the vehicle, such as lateral (steering) and longitudinal (acceleration/braking) behavior. The file also includes my custom auto-centering steering logic. This patch adds a safety automation to re-center the steering wheel when the car is parked and powered off — reducing the risk of accidental adjacent vehicle/property damage, or human injury, if the driver forgets the wheels are still turned as the driver reverses out of a parking space.